backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0

# Database and Storage
supabase==2.7.1

# Vector DB
pinecone-client==2.2.4

# AI/ML dependencies with compatible versions
sentence-transformers==2.7.0
transformers==4.36.2
torch==2.1.2
groq==0.8.0

# LlamaIndex with specific compatible versions
llama-index-core==0.10.57
llama-index-llms-groq==0.1.4
llama-index-vector-stores-pinecone==0.1.3
llama-index-embeddings-huggingface==0.2.0

# PDF processing
PyMuPDF==1.23.26

# Production server
gunicorn==21.2.0
